fix: 마이페이지 수정

+++ src/main/java/org/myteam/server/mypage/dto/response/MyPageResponse.java
@@ -20,7 +20,7 @@ public record MyPageResponse() {
     public static class MemberStatsResponse {
         private int totalVisitCount;       // 전체 방문 횟수
         private int createdPostCount;      // 내가 작성한 게시물 개수
-        private int createdCommentCount;   // 내가 작성한 댓글 개수
+        private long createdCommentCount;   // 내가 작성한 댓글 개수
         private int createdInquiryCount;   // 내가 작성한 문의 개수
         private int totalInviteCount;      // 내가 초대한 횟수
         private String nickname;           // 닉네임
@@ -28,7 +28,7 @@ public record MyPageResponse() {
         private String registeredAt;       // 회원 가입일
         private String registrationMethod; // 가입 방법 (이메일, 소셜 로그인 등)
 
-        public static MemberStatsResponse createResponse(Member member, int postCount, int commentCount, int inquiryCount) {
+        public static MemberStatsResponse createResponse(Member member, int postCount, long commentCount, int inquiryCount) {
             return MemberStatsResponse.builder()
                     .totalVisitCount(member.getMemberActivity().getVisitCount())
                     .createdPostCount(postCount)
